norminette - too many funcs in exec/redirections.c moved to heredoc.c
