+++ source/execution/builtins/cd.c
@@ -75,7 +75,9 @@ int	ft_special_args(t_full_cmd *cmd, t_data *data)
 		ft_setenv("PWD", ft_getenv("HOME", data->env), data->env);
 		return (0);
 	}
-	else if (!ft_strcmp(cmd->args[1], ".") || !ft_strcmp(cmd->args[1], ft_getenv("PWD", data->env)))
+	else if (!ft_strcmp(cmd->args[1], "."))
+		return (0);
+	else if  (!ft_strcmp(cmd->args[1], ft_getenv("PWD", data->env)))
 		return (0);
 	else
 		return (1);
